WIP on splitting IC.Test.Agent and IC.Test.Spec
