+++ ic-hs.cabal
@@ -355,8 +355,10 @@ executable ic-ref-test
   other-modules: IC.Management
   other-modules: IC.Ref.IO
   other-modules: IC.Test.Agent
+  other-modules: IC.Test.Agent.Calls
   other-modules: IC.Test.Options
   other-modules: IC.Test.Spec
+  other-modules: IC.Test.Spec.Utils
   other-modules: IC.Test.Universal
   other-modules: IC.Types
   other-modules: IC.Version
@@ -621,12 +623,14 @@ library
   exposed-modules: IC.Serialise
   exposed-modules: IC.StateFile
   exposed-modules: IC.Test.Agent
+  exposed-modules: IC.Test.Agent.Calls
   exposed-modules: IC.Test.BLS
   exposed-modules: IC.Test.ECDSA
   exposed-modules: IC.Test.HashTree
   exposed-modules: IC.Test.Options
   exposed-modules: IC.Test.Secp256k1
   exposed-modules: IC.Test.Spec
+  exposed-modules: IC.Test.Spec.Utils
   exposed-modules: IC.Test.Universal
   exposed-modules: IC.Test.WebAuthn
   exposed-modules: IC.Types
